rocm-ck: split FMHA BWD headers into spec/api/dev compilation tiers
Introduce a three-file compilation boundary for the FMHA BWD example,
matching the pattern established by the GEMM example (04):

  _spec.hpp — consteval factory, slot constants, structural types
              (compiled in both host and device passes, no HIP dep)
  _api.hpp  — host-only helpers: grid_size calculation
              (#error on __HIP_DEVICE_COMPILE__)
  _dev.hpp  — CK Tile device bridge
              (#error when NOT in device compilation)

Previously, _api.hpp contained both the consteval spec and the
host-only grid_size helper. Because _dev.hpp included _api.hpp, this
pulled <hip/hip_runtime.h> into the device compilation path. The
three-tier split keeps the shared spec HIP-free and enforces the
host/device boundary with #error guards.

Applied uniformly to all three kernel families: OGradDotO, DqDkDv,
and ConvertDQ.
