+++ projects/composablekernel/experimental/rocm_ck/examples/06_rocm_ck_fmha_bwd/rocm_fmha_bwd_dqdkdv_dev.hpp
@@ -9,10 +9,20 @@
 // are host-only, so we initialize directly on the device side.
 //
 // Uses C++20 struct NTTPs: template <FmhaBwdDQDKDVKernel K>.
+//
+// Compilation boundary:
+//   _spec.hpp — consteval factory + slot constants (both passes)
+//   _api.hpp  — host-only helpers: grid_size (host pass only, #error on device)
+//   _dev.hpp (this) — CK Tile bridge + __device__ code (device pass only, #error on host)
 
 #pragma once
 
-#include "rocm_fmha_bwd_dqdkdv_api.hpp"
+#ifndef __HIP_DEVICE_COMPILE__
+#error "rocm_fmha_bwd_dqdkdv_dev.hpp requires device compilation." \
+       " Host code should include rocm_fmha_bwd_dqdkdv_api.hpp."
+#endif
+
+#include "rocm_fmha_bwd_dqdkdv_spec.hpp"
 
 #include <rocm_ck/args.hpp>
 #include <rocm_ck/ck_type_map.hpp>
